Move Money and Scanner to separate files
The Money class was moved to a separate file, since it represents a
significant amount of code as it is. The scanner was moved to a
separate file since it doesn't represent an io operation - rather, it's
parsing a string only.

This commit also fixes the bulk of the non-serious issues with the
code, such as reported spelling errors and misc static checker
warnings.